+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Fix for the case when try to do Settle, Void or Refund on Order made with plugin before v2.0.0.
     * Auto-Void fix. Allow Auto-Void for approved transactions only.
     * Fix for the missing notify URL in create_auto_void() method.
+    * Added ApplePay locale.
 ```
 
 # 2.1.0